+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog #
 
+## 0.1.0 - Feb 18, 2014
+* Added aggregate operations
+** `fold` - collect data.
+** `count` - get number of elements.
+** `pairs` - get key value pairs
+** `keys` - get keys.
+** `values` - get values.
+
 ## 0.0.2 - Feb 18, 2014
 * Removed dead code.
 * Recompiled to remove let expr overhead.